Rename notification handler to plan-upgrade-notification.
