+++ extensions/editor.js
@@ -3,7 +3,7 @@
  */
 import './shared/public-path';
 import './shared/block-category';
-import './shared/notification-on-plan-upgrade';
+import './shared/plan-upgrade-notification';
 import analytics from '../_inc/client/lib/analytics';
 
 // @TODO Please make a shared analytics solution and remove this!
